requirements.txt
streamlit==1.30.0
openai==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
pandas==2.0.0
numpy==1.24.0
folium==0.14.0
Pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.0
geopy==2.4.0
python-dotenv==1.0.0
html5lib==1.1
bleach==6.1.0 